Cleanup and added more optional arguments

+++ mp3-tagger/tagger.py
@@ -4,9 +4,32 @@ from mutagen.easyid3 import EasyID3
 import argparse
 import sys
 
+# Parse args
+parser = argparse.ArgumentParser()
+parser.add_argument("file",
+                    help="specify the file/directory you want to work with")
+parser.add_argument("-d", "--directory",
+                    help="specify that you are providing a directory to be auto tagged",
+                    action="store_true")
+parser.add_argument("-c", "--cleantag",
+                    help="remove current tags",
+                    action="store_true")
+parser.add_argument("-m", "--modifytags",
+                    help="modify existing tags",
+                    action="store_true")
+parser.add_argument("-r", "--rename",
+                    help="rename files after tagging to kebab case",
+                    action="store_true")
+args = parser.parse_args()
+
+
+def create_director_if_does_not_exist(name):
+    if not os.path.exists(name):
+        os.makedirs(name)
+
 
 def get_all_files_in_dir(directory):
-    files = [args.file + x for x in os.listdir(args.file)]
+    files = [os.path.join(args.file, x) for x in os.listdir(args.file)]
     return files
 
 
@@ -21,7 +44,7 @@ def copy_file(directory, filepath):
         copyfile(filepath, outputpath)
         return outputpath
     except Exception as e:
-        print("copying failed! ", e)
+        print("Copying failed! ", e)
         sys.exit(1)
 
 
@@ -48,40 +71,60 @@ def tag(filepath):
     try:
         # Load file and store required information
         audiofile = EasyID3(filepath)
-        currTitle = audiofile["title"]
-        currArtist = audiofile["artist"]
-        currAlbum = audiofile["album"]
+        if not args.cleantag:
+            currTitle = audiofile["title"]
+            currArtist = audiofile["artist"]
+            currAlbum = audiofile["album"]
         # Remove tags
-        audiofile.delete()
-        # Clean tags
-        audiofile["title"] = currTitle
-        audiofile["artist"] = currArtist
-        audiofile["album"] = currAlbum
+        if args.cleantag:
+            audiofile.delete()
+        # Change tags
+        if args.cleantag and args.modifytags:
+            print("New Info ", os.path.basename(filepath))
+            newTitle = input("Title: ")
+            newArtist = input("Artist: ")
+            newAlbum = input("Album: ")
+            audiofile["title"] = newTitle
+            audiofile["artist"] = newArtist
+            audiofile["album"] = newAlbum
+        elif args.modifytags:
+            print("Current info")
+            print("Title: ", currTitle)
+            print("Artist: ", currArtist)
+            print("Album: ", currAlbum)
+            print("New Info")
+            newTitle = input("Title: ")
+            newArtist = input("Artist: ")
+            newAlbum = input("Album: ")
+            audiofile["title"] = newTitle
+            audiofile["artist"] = newArtist
+            audiofile["album"] = newAlbum
         # Save file
         audiofile.save()
         return True
-    except:
+    except Exception as e:
         print("ERROR TAGGING! ", os.path.basename(filepath))
+        print(e)
+        create_director_if_does_not_exist("errors")
         copy_file(os.path.join(os.path.dirname(
-            os.path.realpath(__file__)), "error"), filepath)
+            os.path.realpath(__file__)), "errors"), filepath)
         return False
 
 
-# Parse args
-parser = argparse.ArgumentParser()
-parser.add_argument("file", help="specify the file you want to work with")
-args = parser.parse_args()
-# Program
-
-
-def main():
-    directory = get_all_files_in_dir(args.file)
-    for filepath in directory:
+if __name__ == "__main__":
+    if args.directory:
+        directory = get_all_files_in_dir(args.file)
+        for filepath in directory:
+            create_director_if_does_not_exist("output")
+            x = copy_file(os.path.join(os.path.dirname(
+                os.path.realpath(__file__)), "output"), filepath)
+            if tag(x):
+                if args.rename:
+                    rename_file(x)
+    else:
+        create_director_if_does_not_exist("output")
         x = copy_file(os.path.join(os.path.dirname(
             os.path.realpath(__file__)), "output"), filepath)
         if tag(x):
-            rename_file(x)
-
-
-main()
-print("SUCCESS!")
+            if args.rename:
+                rename_file(x)
